Make -dataLoader:didCancelRequest: an optional delegate method
Most cancellations are done by the API user, so this notification is often
not used to implement any logic, and has most often be left empty since it’s
been a required method.

This commit makes it optional, so that API users that are interested in
this event can react to it, while still removing boilerplate for the most
common use case.