Report downloaded size using Content-Length
Use the Content-Length field of the HTTP response header first.
If missing fallback to the number of bytes reported by the task.
This is useful when compressed data is returned by the server.
In that case, the task will report the uncompressed data instead
of the compressed one that was actually downloaded.