Add a safeguard on NSURLSession invalidation
On a session invalidation, tasks are usually requested to be cancelled.
the `didCompleteWithError` attempts always to create a new task even if
the task had been cancelled. That is usally not an issue unless the
session has been invalidated. In that case a NSGenericException is raied
and the app most likely crashes.